+++ CHANGES.rst
@@ -4,7 +4,8 @@ Products.MeetingLiege Changelog
 4.2.9 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fixed translation of `Data that will be used on new item` on `meetingitem_view.pt`.
+  [gbastien]
 
 
 4.2.8 (2023-03-20)